+++ lib/gamefic-standard/give.rb
@@ -1,19 +1,20 @@
-# @gamefic.script standard/give
-
-respond :give, Use.available, Gamefic::Query::Children.new do |actor, _character, gift|
-  actor.tell "Nothing happens."
-end
-
-respond :give, Use.available(Character), Use.available do |actor, character, gift|
-  actor.tell "#{The character} doesn't want #{the gift}."
-end
-
-respond :give, Use.available(Character), Use.available do |actor, _character, gift|
-  if gift.parent == actor
-    actor.proceed
-  else
-    actor.tell "You don't have #{the gift}."
+Gamefic.script do
+  respond :give, Use.available, Gamefic::Query::Children.new do |actor, _character, _gift|
+    actor.tell 'Nothing happens.'
   end
-end
 
-interpret "give :gift to :character", "give :character :gift"
+  respond :give, Use.available(Character), Use.available do |actor, character, gift|
+    actor.tell "#{The character} doesn't want #{the gift}."
+  end
+
+  respond :give, Use.available(Character), Use.available do |actor, _character, gift|
+    if gift.parent != actor
+      actor.perform :take, gift
+    end
+    if gift.parent == actor
+      actor.proceed
+    end
+  end
+
+  interpret 'give :gift to :character', 'give :character :gift'
+end