+++ lib/gamefic-standard/entities/character.rb
@@ -1,11 +1,2 @@
-class Character < Thing
-  include Gamefic::Active
-
-  def gender
-    @gender ||= :other
-  end
-end
-
-Gamefic.script do
-  set_player_class Character
-end
+Character = Gamefic::Actor
+Character.set_default gender: :other