+++ lib/gamefic-standard/modules.rb
@@ -3,6 +3,7 @@
 # portals between rooms.
 #
 
+require 'gamefic-standard/modules/standardized'
 require 'gamefic-standard/modules/use'
 require 'gamefic-standard/modules/enterable'
 require 'gamefic-standard/modules/openable'