+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 2.3.0
 - Use nil actions for untokenized arguments
+- Fix response to closing closed entities
 
 ## 2.2.0 - June 21, 2021
 - Insert, place, and wear try take first