+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 2.3.0
 - Use nil actions for untokenized arguments
 - Fix response to closing closed entities
+- Working give action
 
 ## 2.2.0 - June 21, 2021
 - Insert, place, and wear try take first