+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - Use nil actions for untokenized arguments
 - Fix response to closing closed entities
 - Working give action
+- Miscellaneous fixes for new specs
 
 ## 2.2.0 - June 21, 2021
 - Insert, place, and wear try take first